% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve_vectorized.R
\name{add_peak_info}
\alias{add_peak_info}
\title{add_peak_info}
\usage{
add_peak_info(x, gam, lambda, constraint, estimate_calcium, var = var, ESP)
}
\arguments{
\item{ESP}{}
}
\description{
add_peak_info
}
