% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/alignEpocs.R
\name{alignEpocs}
\alias{alignEpocs}
\title{alignEpocs}
\usage{
alignEpocs(behavior_data, fiber_data, mouse_id)
}
\arguments{
\item{behavior_data}{A data table containing behavioral annotations timestamped
with the annotateVideo function}

\item{fiber_data}{A data table containing calcium data, isosbestic data,
mouse ID and the time (in seconds) as generated with the
extract_tdt_data function or extract_all_tdt_data function}

\item{mouse_id}{A character indicating the mouse ID.

This function can be use with a single video annotation file (one mouse) and
a dataset containing all fiber photometry data (all mice). Or with a single
video annotation file and a fiber photometry data containing only this mouse.
To process all video annotations and all fiber photometry data, use this function
iteratively with apply or a for loop.}
}
\value{
A joint data table with behavioral data aligned to fiber photometry data
}
\description{
alignEpocs
}
