% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Responses_vectorized.R
\name{Analyze_Responses}
\alias{Analyze_Responses}
\title{Analyze_Responses}
\usage{
Analyze_Responses(
  data,
  df_clean,
  compare_groups = FALSE,
  one_cell = FALSE,
  marker = FALSE,
  var_list = NULL,
  simulation = FALSE
)
}
\arguments{
\item{compare_groups}{}
}
\description{
Analyze_Responses
}
