% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base_resp.rm.R
\name{base_resp.rm}
\alias{base_resp.rm}
\title{base_resp.rm}
\usage{
base_resp.rm(borders, full)
}
\arguments{
\item{borders}{}
}
\value{
The dataset containing the peaks informations without the peaks of
the cells that responded to baseline.
}
\description{
base_resp.rm
}
