% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions_CalQuick.R
\name{cell_plot}
\alias{cell_plot}
\title{cell_plot}
\usage{
cell_plot(
  full_data,
  peaks_data,
  cell,
  var,
  line = c(FALSE, "poly", "gam", "background"),
  show_peak = FALSE
)
}
\arguments{
\item{show_peak}{}
}
\description{
cell_plot
}
