% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_Sort_CalQuick.R
\name{cell_sort}
\alias{cell_sort}
\title{cell_sort}
\usage{
cell_sort(
  df,
  pat,
  duration_in_seconds,
  frame_rate,
  id,
  multiple = TRUE,
  compare_groups = TRUE,
  groups
)
}
\arguments{
\item{groups}{}
}
\description{
cell_sort
}
