% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareClustData.R
\name{clustCellID}
\alias{clustCellID}
\title{clustCellID}
\usage{
clustCellID(clust_res, peaks, to_rm)
}
\arguments{
\item{clust_res}{dtwclust object resulting from a clustering step}

\item{peaks}{data table containing peaks in lines}

\item{to_rm}{a vector containing cell names (e.g. "A1aaa) to exclude from the identification. Leave empty keeping all cells.}
}
\value{
the peaks_data data table annotated with a new variable "cluster"
}
\description{
Labels each cell in peaks_data with the corresponding cluster.
}
