% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{CN_DPA}
\alias{CN_DPA}
\title{CN_DPA}
\usage{
CN_DPA(data, CN_DPA_width)
}
\arguments{
\item{CN_DPA_width}{}
}
\description{
Implementation of a constrained non negative variant of the derivative passing accumulation method.
The persistence of the "potential signal" can be adapted by lagging the initial positive vector
}
