% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Responses.R, R/Analyze_responses_bis.R
\name{Compare_props}
\alias{Compare_props}
\title{Compare_props}
\usage{
Compare_props(data)

Compare_props(data)
}
\arguments{
\item{data}{the full output from Count_responders_stim()}
}
\value{
a list with 2 elements. The first is the general summary of the cochran's q test
the second is the results of the pairwise comparisons between stimuli with McNemar test with correction for multiple comparisons
}
\description{
Compare_props

Compare_props
}
