% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_responses_bis.R
\name{Count_responders}
\alias{Count_responders}
\title{Count_responders}
\usage{
Count_responders(data, df_tidy)
}
\arguments{
\item{data}{the first data frame output from find_rise or keep_best_peaks. It must be a data frame
with each row describing one response}

\item{df_tidy}{the full data frame after cleaning (the output from clean_data)}
}
\value{
the number of cells that responded at least once and the proportion of responses it corresponds to
}
\description{
Count_responders
}
