% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_responses_bis.R
\name{Count_responders_stim}
\alias{Count_responders_stim}
\title{Count_responders_stim}
\usage{
Count_responders_stim(data, df_tidy, n_cells)
}
\arguments{
\item{data}{the first data frame output from find_rise}

\item{df_tidy}{the full but cleaned data frame (output from clean_data())}
}
\value{
a table summarizing the number of peaks for each stimulus (will be used in compare_props function)
and a data frame gathering the number/proportion of cells that responded to each stimulus
}
\description{
Count_responders_stim
}
