% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CreateId.R
\name{createId}
\alias{createId}
\title{createId}
\usage{
createId(df, coverslip_id)
}
\arguments{
\item{coverslip_id}{}
}
\description{
Creates a unique identifier for each Cell which keeps the original
cell order in the raw csv file. The identifier contains the coverslip
of the cells.
}
