% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dual_prop.R, R/dual_prop_CalQuick.R
\name{cross_prop.venn}
\alias{cross_prop.venn}
\title{cross_prop.venn}
\usage{
cross_prop.venn(data, stim_list)

cross_prop.venn(data, stim_list)
}
\arguments{
\item{stim2}{}
}
\description{
cross_prop.venn

cross_prop.venn
}
