% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dashboardHeader.R
\name{dashboardHeader}
\alias{dashboardHeader}
\title{dashboardHeader}
\usage{
dashboardHeader(
  ...,
  title = NULL,
  disable = FALSE,
  title.navbar = NULL,
  .list = NULL
)
}
\arguments{
\item{.list}{}
}
\description{
A variant of the dashboardHeader function from shinydashboard package that
allows to add a navbar title.
Taken from : https://github.com/rstudio/shinydashboard/issues/41
}
