% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve_CalQuick.R
\name{deconvolve}
\alias{deconvolve}
\title{deconvolve}
\usage{
deconvolve(
  norm_data,
  gam = 0.95,
  lambda = 1,
  constraint = T,
  estimate_calcium = T,
  var = "smooth_delta",
  ESP = 1e-04,
  threshold = 3
)
}
\arguments{
\item{threshold}{}
}
\description{
deconvolve
}
