% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/delta_f_CalQuick.R
\name{delta_f}
\alias{delta_f}
\title{delta_f}
\usage{
delta_f(cov_split, var = c("poly", "gam", "linear", "quantile"))
}
\arguments{
\item{cov_split}{a list of data table, each element being one coverslip (one technical replicate)}

\item{var}{the variable which will be normalized (one of poly, gam, linear, quantile)}

\item{data}{a dataframe output from back_estimate function}
}
\value{
a vector of the normalized fluorescence values with the delta f / f
method, taking the predicted background as a local baseline
}
\description{
delta_f
}
