% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_raw_CalQuick.R
\name{derive_raw}
\alias{derive_raw}
\title{derive_raw}
\usage{
derive_raw(x)
}
\arguments{
\item{x}{A dataframe output from back estimate function at least, not earlier in the process}
}
\description{
This function computes the first derivative on the raw detrended fluorescence values.
}
