% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derive_smooth_CalQuick.R
\name{derive_smooth}
\alias{derive_smooth}
\title{derive_smooth}
\usage{
derive_smooth(x)
}
\arguments{
\item{x}{}
}
\description{
This function computes the first derivative on the standardized + smoothed fluorescence values
}
