% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backEstimate_pat_match.R
\name{distcomputR}
\alias{distcomputR}
\title{distcomputR_fast}
\usage{
distcomputR(patMat, subseqMat, step, window)
}
\arguments{
\item{window}{}
}
\description{
Computing and extracting the distance between each subsequence and
}
