% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{downsampleCaData}
\alias{downsampleCaData}
\title{downsampleCaData}
\usage{
downsampleCaData(data, original_freq, target_freq)
}
\arguments{
\item{data}{data table output from prepareData function}

\item{original_freq}{Frequency of the data argument (in Hz)}

\item{target_freq}{Desired Frequency after downsampling (in Hz)}
}
\value{
a data table downsampled at the desired frequency
}
\description{
downsampleCaData
}
