% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periEvents.R
\name{eventExtractR}
\alias{eventExtractR}
\title{eventExtractR}
\usage{
eventExtractR(data)
}
\arguments{
\item{data}{A data table containing aligned behavioral and fiber photometry data
(e.g. the output of alignEpocs())}
}
\description{
This function extracts all the behavioral events onsets
}
