% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakExtractR.R
\name{extractPeakVal}
\alias{extractPeakVal}
\title{extractPeakVal}
\usage{
extractPeakVal(full_data, peaks_data)
}
\arguments{
\item{full_data}{second element from the list returned by peakExtractR function}

\item{peaks_data}{first element from the list returned by peakExtractR function}
}
\value{
a data table containing the whole fluorescence values between the start
and the end of each detected peak.
}
\description{
Extracts the whole time series (from start to end) of each detected peaks.
This is used to then compute the area under the curve (AUC) of each peak.
}
