% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivativeRatio_CalQuick.R,
%   R/derivativeRatio_CalQuick_nouvelle_version.R
\name{false_pos}
\alias{false_pos}
\title{derivativeRatio}
\usage{
false_pos(peaks_data, full_data, range)

false_pos(peaks_data, full_data, range)
}
\arguments{
\item{range}{}
}
\description{
derivativeRatio

derivativeRatio
}
