% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_baseline_response_CalQuick.R
\name{filter_baseline_response}
\alias{filter_baseline_response}
\title{filter_baseline_response}
\usage{
filter_baseline_response(data, threshold)
}
\arguments{
\item{threshold}{}
}
\value{
a filtered dataframe with cells responding at baseline removed
}
\description{
filter_baseline_response
}
