% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_borders_CalQuick.R
\name{find_borders}
\alias{find_borders}
\title{find_borders}
\usage{
find_borders(data, range)
}
\arguments{
\item{range}{}
}
\description{
This function finds the start and the end of each peak that has been found
from the previous step
}
