% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_end.R
\name{find_end}
\alias{find_end}
\title{find_end}
\usage{
find_end(peaks, full_data, range)
}
\arguments{
\item{peaks}{A list of dataframe with each element being the informations of a peak}

\item{range}{the time window after the maxima of the peak in wich to look for the end}
}
\value{
A list of data table objects with each element being the informations regarding the end of one peak
}
\description{
This function finds the end of each peak in the dataset. It is based on
the trapezoid method as described in :
https://www.researchgate.net/profile/Roberto-Oliveira-7/publication/51637823_New_approach_for_T-wave_end_detection_on_electrocardiogram_Performance_in_noisy_conditions/links/004635204e355e4bed000000/New-approach-for-T-wave-end-detection-on-electrocardiogram-Performance-in-noisy-conditions.pdf?_sg\%5B0\%5D=e63wPE5M2qk78DidbANkTHjPauBikqoopZck3Hrj9eqqA7w5KrtUB6sNLfqSW9SW7ZYgaWkL9pb9kBdKmhL0mg.3bSqx9fAuEE_4YmwlOnGgFtq-7uSsouXgfLVmhw0abvkwJGoWvnjWJI6TSHCNR_tl-MXl4b4xIQOsHC3nsQtTQ&_sg\%5B1\%5D=xG1pCg_18Q1boMc0hcDjNAb7_J1kIvrvE1BJ9Wzfq0xiVLYXGhpICPqjvK2tv0LLi7FtXzbwchBqH9y2zpLFyWRMC-SqMuU7UT2NTqwL7MwE.3bSqx9fAuEE_4YmwlOnGgFtq-7uSsouXgfLVmhw0abvkwJGoWvnjWJI6TSHCNR_tl-MXl4b4xIQOsHC3nsQtTQ&_sg\%5B2\%5D=okeeSpIP3PH3D6113upLcK4TWZ5hLfbDzb-p2-Cf-VnfnszQIU3SXbz9kQ-v188m9_jCJq74Vrw1SuI.R-cwcRofo6lYn_UKG_b_vpgEaw3REI89B7mUiQWIOH_cTGwL6nHAXMwO5ZTDD73QR1ToSWp9OmiTl8W4LBPdHA&_iepl=
}
