% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_rise_CalQuick.R,
%   R/find_rise_bis_CalQuick.R
\name{find_rise}
\alias{find_rise}
\title{find_rise}
\usage{
find_rise(data, range)

find_rise(data, range)
}
\arguments{
\item{range}{}
}
\description{
find_rise

find_rise
}
