% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backEstimate_pat_match.R
\name{interpolR}
\alias{interpolR}
\title{interpolR}
\usage{
interpolR(list, len, type = c("one", "multiple"))
}
\arguments{
\item{len}{}
}
\value{
a matrix where each initial pattern has length len.
}
\description{
Takes a list of time series patterns and the length to which each must be
linearly interpolated
}
