% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keep_best_peaks.R
\name{keep_best_peaks}
\alias{keep_best_peaks}
\title{keep_best_peaks}
\usage{
keep_best_peaks(data)
}
\arguments{
\item{data}{the whole output from find_peaks function}

\item{range}{}
}
\value{
the same objects with an added column named True_Peaks which is TRUE if the peak is kept, FALSE if not
}
\description{
keep_best_peaks
}
