% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_GuiGui.R
\name{launch_GuiGui_bis}
\alias{launch_GuiGui_bis}
\title{launch_GuiGui}
\usage{
launch_GuiGui_bis()
}
\description{
Exectue this function without any argument to launch the Graphical User Interface (GUI)
}
