% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backEstimate_pat_match.R
\name{matFillR_bis_bis}
\alias{matFillR_bis_bis}
\title{matFillR_bis_bis}
\usage{
matFillR_bis_bis(dt)
}
\arguments{
\item{dt}{}
}
\description{
Creates a matrix to align all the distance values obtained with the different
subsequences together with the indexes of the original trace. For each index,
the kept distance is the median across all the overlapping distances
}
