% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backEstimate_pat_match.R
\name{mdRatio}
\alias{mdRatio}
\title{mdRatio}
\usage{
mdRatio(pos, neg)
}
\arguments{
\item{neg}{}
}
\value{
a data table with the median ratio in one column
}
\description{
Computes the median distance ratio between the best match in the positive and
the negative banks, for each index.
}
