% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_en_travaux_CalQuick.R
\name{moving_cells}
\alias{moving_cells}
\title{moving_cells}
\usage{
moving_cells(data, threshold = 0.1)
}
\arguments{
\item{threshold}{An optional parameter defining over how many standard deviations (of the baseline)
a decrease in median during a stimulus will be considered a moving cell and thus discarded.}
}
\value{
a data frame with all moving cells removed
}
\description{
moving_cells
}
