% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data.R
\name{moving_cells_old}
\alias{moving_cells_old}
\title{moving_cells}
\usage{
moving_cells_old(data, threshold = 0.1)
}
\arguments{
\item{threshold}{An optional parameter defining over how many standard deviations (of the baseline)
a decrease in median during a stimulus will be considered a moving cell and thus discarded.}
}
\value{
a data frame with all moving cells removed
}
\description{
moving_cells
}
