% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{norm_df}
\alias{norm_df}
\title{norm_df}
\usage{
norm_df(
  data,
  var = c("raw", "poly", "gam", "linear", "quantile", "back"),
  width,
  reference = c("baseline", "estimate")
)
}
\arguments{
\item{data}{a data frame output from prepareData and back_estimate}

\item{cov_split}{A list of data frame with each element being the data for one coverslip (one technical replicate)}

\item{dim_list}{A list with each element being the dimension of a data frame in cov_split}
}
\value{
An updated data frame with four new columns : z_score (the values transformed in z_score)
and the smoothed version of the z_score ; delta_f_f (the values normalized by delta f /f method)
and its smoothed version
}
\description{
A wrapper which executes z score noralization, deltaf/f normalization
and computes a smoothed version of it
}
