% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backEstimate_pat_match.R
\name{patDetectR}
\alias{patDetectR}
\title{patDetectR}
\usage{
patDetectR(dt, window, step, posBank, negBank, new_len, Var, Norm = TRUE)
}
\arguments{
\item{Norm}{}
}
\value{
A data table containing the ratio for each index and related informations
}
\description{
Computes the dynamic time warping distance between all the rolling subsequences
of a given length (or several given lengths) and all the patterns in two pattern
banks: a positive bank and a negative bank (and returns the distance for the
best match in each bank).So each index end up with two distance values for each
specified length (window). The value couple (pos;neg)
containing the value that is minimal, is kept. This value couple is then
used to compute a ratio (pos/neg), used to infer if the index belongs to signal
or noise
}
