% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakExtractR.R
\name{peakExtraction}
\alias{peakExtraction}
\title{peakExtraction}
\usage{
peakExtraction(deconvolved, thresh = 3, delta_thresh = 0, var, peak_frame)
}
\arguments{
\item{deconvolved}{the object output from the deconvolve function}

\item{thresh}{a numeric indicating the z score threshold a peak maximum must
exceed (or be equal to) to be kept, default to 3 (i.e. 3 standard deviations)}

\item{delta_thresh}{a numeric indicating the delta f/f threshold a peak maximum must
exceed (or be equal to) to be kept default to 0}

\item{var}{the variable used for deconvolution (default to "background_detrended").
If you don't use the pattern matching option for noise estimation and use DPA method instead,
this parameter needs to be changed to "gam_detrended"}

\item{peak_frame}{an integer corresponding to the number of frame you want
to screen after each spike to find the peak maximum default to 10 (tested in 1 and 0.25Hz data)}
}
\value{
a list with two data tables. The first is the peaks data as in the
output of deconvolve but with new columns indicating the beginning and end of each event
the maximum and the AUC of the overall event. The second is the full data as in the
output of deconvolve.
}
\description{
This function takes the output from the deconvolve function to extract
full calcium events (peaks), attributing only one start and one end to each event.
It then computes the area under the curve (AUC) of each event.
}
