% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peakExtractR.R
\name{peakExtractR}
\alias{peakExtractR}
\title{peakExtractR}
\usage{
peakExtractR(
  peaks_data,
  norm_data,
  peak_frame = 10,
  threshold = 3,
  delta_threshold = 0,
  var = "background_detrended"
)
}
\arguments{
\item{peaks_data}{a data table with each line corresponding to a spike found
during the deconvolution. This is the first element of the list returned by
the deconvolve() function}

\item{norm_data}{a data table output from the norm_df() function}

\item{peak_frame}{an integer corresponding to the number of frame you want
to screen after each spike to find the peak maximum default to 10 (tested in 1 and 0.25Hz data)}

\item{threshold}{a numeric indicating the z score threshold a peak maximum must
exceed (or be equal to) to be kept, default to 3 (i.e. 3 standard deviations)}

\item{delta_threshold}{a numeric indicating the delta f/f threshold a peak maximum must
exceed (or be equal to) to be kept default to 0}

\item{var}{the variable used for deconvolution (default to "background_detrended").
If you don't use the pattern matching option for noise estimation and use DPA method instead,
this parameter needs to be changed to "gam_detrended"}
}
\value{
a list with 3 elements : a data table with peaks data (with peak start and end estimated),
the norm data and a logical indicating if peaks have been found.
}
\description{
peakExtractR
}
