% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/periEvents.R
\name{PeriEventExtractR}
\alias{PeriEventExtractR}
\title{PeriEventExtractR}
\usage{
PeriEventExtractR(data, before, after)
}
\arguments{
\item{data}{A data table with behavioral annotations and photometry data
aligned together, as returned by the alignEpocs function.}

\item{before}{A numeric indicating how many seconds before stimulus onset should be kept.}

\item{after}{A numeric indicating how many seconds after stimulus onset should be kept.}
}
\value{
A list of data tables with two levels. The first level is the same length
as the number of different behavioral events in the dataset. For each event, a list
of data table is stored. Each data table contains photometry and behavioral informations
10 seconds before the onset of the event until 10 seconds after the onset.
}
\description{
PeriEventExtractR
}
