% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_aligned_fiber_data}
\alias{plot_aligned_fiber_data}
\title{plot_aligned_fiber_data}
\usage{
plot_aligned_fiber_data(
  data,
  x_var,
  y_var,
  isos = FALSE,
  norm = FALSE,
  behavior = FALSE
)
}
\arguments{
\item{data}{a data.table object containing x_var, y_var, the isosbestic trace fitted to the raw trace (fit_isos)
and the normalized calcium trace (delta_f_f), all in separated columns}

\item{x_var}{a numeric vector containing the time}

\item{y_var}{a numeric vector containing the calcium trace (or any other information)}

\item{isos}{a logical indicating if isosbestic trace should be plotted}

\item{norm}{a logical indicating if the normalized trace (deltaf/f) should be
plotted instead of the raw trace.}
}
\description{
plot_aligned_fiber_data
}
