% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{plot_fiber_data}
\alias{plot_fiber_data}
\title{plot_fiber_data}
\usage{
plot_fiber_data(data, x_var, y_var)
}
\arguments{
\item{data}{A data.table object containing x_var and y_var in 2 separate columns}

\item{x_var}{a numeric vector containing the time}

\item{y_var}{a numeric vector containing the calcium trace (or any other information)}
}
\description{
plot_fiber_data
}
