% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/derivativeRatio_CalQuick.R,
%   R/derivativeRatio_CalQuick_nouvelle_version.R
\name{PNHRA}
\alias{PNHRA}
\title{derivativeRatio}
\usage{
PNHRA(peaks_data, full_data, range)

PNHRA(peaks_data, full_data, range)
}
\arguments{
\item{range}{}
}
\description{
derivativeRatio

derivativeRatio
}
