% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{prepareData}
\alias{prepareData}
\title{prepareData}
\usage{
prepareData(
  folder_name,
  stim_number,
  frame_rate,
  duration_in_seconds = 30,
  compare_groups = FALSE,
  marker_thresh
)
}
\arguments{
\item{compare_groups}{}
}
\description{
This function is a wrapper which reads the raw fluorescence csv files (wide format),
transform it to long format and adds several meta in column (Cell_id, time, stimuli,
molecular marker, group, coverslip).
}
