% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{prepareData_track}
\alias{prepareData_track}
\title{prepareData_track}
\usage{
prepareData_track(
  folder_name,
  frame_rate,
  target_rate = 0,
  duration_in_seconds = 30,
  marker_thresh = 0
)
}
\arguments{
\item{compare_groups}{}
}
\description{
prepareData_track
}
