% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Response_by_stim_and_groups.R
\name{Response_by_stim_and_groups}
\alias{Response_by_stim_and_groups}
\title{Response_by_stim_and_groups}
\usage{
Response_by_stim_and_groups(data, df_clean)
}
\arguments{
\item{data}{the first output from find_rise() (a data frame with one line summarizing the informations of one response)}

\item{df_clean}{the output from clean_data() (the full but cleaned data frame)}
}
\description{
Response_by_stim_and_groups
}
