% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_calipR_env.R
\name{set_calipR_env}
\alias{set_calipR_env}
\title{set_calipR_env}
\usage{
set_calipR_env()
}
\description{
This function is used to set the current environment to the portable conda
environment included in calipR
}
