% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions_CalQuick.R
\name{simple_cell_plot}
\alias{simple_cell_plot}
\title{simple_cell_plot}
\usage{
simple_cell_plot(
  data,
  triangle_data,
  cell,
  var,
  line = c(FALSE, "poly", "gam", "quantile"),
  trianglePlot = FALSE
)
}
\arguments{
\item{line}{}
}
\description{
simple_cell_plot
}
