% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{stim_var}
\alias{stim_var}
\title{stim_var}
\usage{
stim_var(data, stimuli, each, frame_rate, coverslip_id)
}
\arguments{
\item{data}{the original dataset}

\item{stimuli}{a character vector with the names of the stimuli used}

\item{each}{an integer vector with the number of repetitions for each stimulus}
}
\value{
Returns a dataframe with a new variable called stimulus and coverslip
}
\description{
Takes the wide format data, the meta.csv file and the coverslip names. For each
data table (each file), it creates a stimulus variable which duplicates the name
of each stimulus the adequate number of times (depending on the time during which
each stimulus was applied), which is specified in the meta file.The coverslip
id is also replicated to match the number of line in each data table.
}
