% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backEstimate_pat_match.R
\name{subinoR}
\alias{subinoR}
\title{subinoR}
\usage{
subinoR(
  dt,
  window,
  step,
  new_len,
  posBank,
  negBank,
  norm = TRUE,
  var = "Mean_Grey"
)
}
\arguments{
\item{var}{}
}
\value{
a list containing a data table with all the interpolated and
normalized subsequences for each trace, two data tables with the interpolated
and normalized positive and negative banks and a list with the number of subsequences
for each specified subsequence window.
}
\description{
Wrapper which does three preparation steps before computing the DTW distance
between the banks: subsequencing, interpolation, z normalization
}
