% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backEstimate_pat_match.R
\name{subsequencR}
\alias{subsequencR}
\title{subsequencR}
\usage{
subsequencR(time_series, window, step, resizing)
}
\arguments{
\item{step}{}
}
\value{
a data table of subsequences
}
\description{
Vectorized function to subset equal length windows from a longer time series.
}
