% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions_CalQuick.R
\name{svg_plot}
\alias{svg_plot}
\title{svg_plot}
\usage{
svg_plot(
  full_data,
  peaks_data,
  cell,
  var,
  line = c(FALSE, "poly", "gam"),
  show_peak = FALSE,
  file_name,
  width,
  height
)
}
\arguments{
\item{height}{}
}
\description{
svg_plot
}
