% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting_functions.R
\name{svg_responders}
\alias{svg_responders}
\title{svg_responders}
\usage{
svg_responders(
  full_data,
  peaks_data,
  var,
  line = c(FALSE, "poly", "gam"),
  show_peak = FALSE,
  width,
  height
)
}
\arguments{
\item{height}{}
}
\description{
svg_responders
}
