% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{tidy_df}
\alias{tidy_df}
\title{tidy_df}
\usage{
tidy_df(
  data,
  stimuli,
  each,
  pattern,
  duration_in_seconds,
  frame_rate,
  coverslip_id,
  id,
  multiple = FALSE,
  groups,
  marker,
  marker_thresh,
  unit
)
}
\arguments{
\item{data}{the original raw dataset}

\item{stimuli}{a character vector with the names of the stimuli used}

\item{each}{an integer vector with the number of repetitions for each stimulus}

\item{pattern}{a character pattern in the colnames of the cells to be used
to gather the cells in one column}
}
\value{
A fully tidied dataset ready for filtering step
}
\description{
tidy_df
}
