% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{time_to_f}
\alias{time_to_f}
\title{time_to_f}
\usage{
time_to_f(x, frame_rate, unit = c("seconds", "minutes"))
}
\arguments{
\item{unit}{}
}
\description{
Converts time in minutes or seconds in frame
}
