% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zScore_CalQuick.R
\name{z_score}
\alias{z_score}
\title{z_score}
\usage{
z_score(
  data,
  var = c("raw", "poly", "gam", "linear", "quantile"),
  cov_split,
  dim_list
)
}
\arguments{
\item{data}{The output of prepareData function or a data frame with similar elements}

\item{cov_split}{A list of data frame with each element being the data for one coverslip (one technical replicate)}

\item{dim_list}{}

\item{raw}{An optional logical parameter. If TRUE, The z score is computed on raw data}

\item{detrended}{An optional logical parameter. If TRUE, The z score is computed on detrended data}

\item{poly_std}{An optional logical parameter. If TRUE, The z score is computed on data normalized through polynomial fit}
}
\value{
the normalized values (z score) as a vector of floating numbers in the same order of the original data frame
}
\description{
z_score
}
