% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Responses_vectorized.R
\name{Analyze_Responses}
\alias{Analyze_Responses}
\title{Analyze_Responses}
\usage{
Analyze_Responses(
  data,
  df_clean,
  compare_groups = FALSE,
  one_cell = FALSE,
  marker = FALSE,
  var_list = NULL,
  simulation = FALSE,
  by = NULL,
  var = NULL,
  stims,
  alt = alt
)
}
\arguments{
\item{compare_groups}{}
}
\description{
Takes the full data table after the cleaning step and the peaks data table
(one line per calcium event) to provide counts and proportions of responding cells as a function of
a (combination of) user defined variable(s). If values are repeated measures, a cochran q test evaluates difference in at
least one condition then post hoc pairwise comparisons are calculated with a mcnemar test.
}
