% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Analyze_Responses_vectorized.R
\name{base_resp.rm}
\alias{base_resp.rm}
\title{base_resp.rm}
\usage{
base_resp.rm(borders, full)
}
\arguments{
\item{borders}{}
}
\value{
each dataset without data of cells responding to baseline
}
\description{
Identifies and removes baseline responders from the full and the peaks datatable
}
