% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Cell_Sort.R
\name{cell_sort}
\alias{cell_sort}
\title{cell_sort}
\usage{
cell_sort(
  df,
  pat,
  duration_in_seconds,
  frame_rate,
  id,
  multiple = TRUE,
  compare_groups = TRUE,
  groups,
  marker,
  marker_thresh
)
}
\arguments{
\item{marker_thresh}{}
}
\value{
A long format data table with 8 columns : Mean_Grey, Cell_id, stimulus, coverslip, time_frame,
time_seconds, time_frame_stim, marker, marker_positive
}
\description{
Transforms the raw datatable in long format, creates a unique Identifier for each cell
and adds coverslip, timing, grouping and molecular marker informations.
}
