% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cleaning.R
\name{clean_data}
\alias{clean_data}
\title{clean_data}
\usage{
clean_data(
  data,
  moving_threshold,
  outlier_threshold,
  mean_width,
  CN_DPA_width,
  DPA_width,
  mean_width_diff,
  method = "DPA"
)
}
\arguments{
\item{data}{a data.table output from prepareData function}
}
\description{
This function removes cells that are moving, dying, not responding or
with mean fluorescence values that are deviating more than n sd from the
population's sd
}
