% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_data_en_travaux_CalQuick.R
\name{clean_data}
\alias{clean_data}
\title{clean_data}
\usage{
clean_data(
  data,
  moving_threshold,
  outlier_threshold,
  mean_width,
  DPA_width,
  mean_width_diff,
  mean_width_second
)
}
\arguments{
\item{data}{a data.table output from prepareData function}

\item{threshold}{a number of sd above which : a cell's mean fluorescence is
considered an outlier, an ending median fluorescence is too high compared with
the minimum median fluorescence}

\item{width}{the number of frames that will be taken to compute the
running average (this is used to smooth the data to then fit a linear model
and exclude cells with a fit (adjusted R2) above 80\% )}
}
\description{
This function removes cells that are moving, dying, not responding or
with mean fluorescence values that are deviating more than n sd from the
population's sd
}
