% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CN_DPA.R
\name{CN_DPA}
\alias{CN_DPA}
\title{CN_DPA}
\usage{
CN_DPA(data, CN_DPA_width)
}
\arguments{
\item{CN_DPA_width}{}
}
\description{
Implementation of a constrained non negative variant of the derivative passing accumulation method
}
