% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolve.R, R/deconvolve_vectorized.R
\name{deconvolve}
\alias{deconvolve}
\title{deconvolve}
\usage{
deconvolve(
  norm_data,
  gam = 0.95,
  lambda = 1,
  constraint = T,
  estimate_calcium = T,
  var = "gam_detrended",
  ESP = 1e-04,
  threshold = 3,
  delta_threshold = 0,
  peak_frame = 10
)

deconvolve(
  norm_data,
  gam = 0.95,
  lambda = 1,
  constraint = T,
  estimate_calcium = T,
  var = "gam_detrended",
  ESP = 1e-04,
  threshold = 3,
  delta_threshold = 0,
  peak_frame = 10
)
}
\arguments{
\item{threshold}{}
}
\description{
deconvolve

deconvolve
}
