% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deconvolution.R
\name{deconvolve}
\alias{deconvolve}
\title{deconvolve}
\usage{
deconvolve(
  norm_data,
  gam = 0.95,
  lambda = 1,
  constraint = T,
  estimate_calcium = T,
  var = "gam_detrended",
  ESP = 1e-04,
  threshold = 3,
  delta_threshold = 0,
  peak_frame = 10
)
}
\arguments{
\item{threshold}{}
}
\description{
Deconvolution of the denoised fluorescence trace based on the FastL0SpikeInference package.
To date, this step is used for two things : identifiying the start of calcium events and efficiently identifying
multiple peaks during a given stimulus. Found spikes are not used as is to infer spiking activity but will be in the future.
}
