% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/normalization.R
\name{delta_f}
\alias{delta_f}
\title{delta_f}
\usage{
delta_f(cov_split, var = c("poly", "gam", "linear", "quantile", "back"))
}
\arguments{
\item{cov_split}{a list of data table, each element being one coverslip (one technical replicate)}

\item{var}{the variable which will be normalized (one of poly, gam, linear, quantile)}

\item{data}{a dataframe output from back_estimate function}
}
\value{
the cov_split list within which, each datatable has a new column : the normalized fluorescence values
}
\description{
Normalizes the denoised fluorescence trace with the delta f/f method taking the predicted background as a local baseline
}
