% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/downstream_analysis.R
\name{downstream_analysis}
\alias{downstream_analysis}
\title{downstream_analysis}
\usage{
downstream_analysis(
  data,
  moving_thresh = 0.1,
  outlier_thresh = 2,
  mean_width = 20,
  DPA_width = 10,
  CN_DPA_width = 20,
  mean_width_diff = 10,
  method = "gam",
  norm_var = "gam",
  norm_width = 10,
  lambda = 100,
  gam = 0.97,
  constraint = T,
  z_thresh = 3,
  delta_thresh = 0,
  deconvolve_var = "gam_detrended",
  borders_range = 50,
  time_thresh = 1,
  compare_groups = FALSE,
  false_pos = c(TRUE, FALSE),
  one_cell = FALSE,
  simulation = FALSE,
  pattern_matching = FALSE,
  posBank = list(),
  negBank = list(),
  windows = c(30, 70, 100)
)
}
\arguments{
\item{steps}{}
}
\description{
A wrapper function used in the graphical user interface to process the already prepared data.
Can be used as a one line function to run the complete workflow on the output from prepareData or prepareata_track.
}
