% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/false_positive_estimation.R
\name{keep_best_peaks}
\alias{keep_best_peaks}
\title{keep_best_peaks}
\usage{
keep_best_peaks(data)
}
\arguments{
\item{data}{the whole output from find_peaks function}

\item{range}{}
}
\value{
the same objects with an added column named True_Peaks which is TRUE if the peak is kept, FALSE if not
}
\description{
This function is optional in the workflow. It tries to identify false positives
to remove them. It is based on two parameters : (1) the number of value pairs
(-1 and +1 around the peak, -2 and +2 around the peak) that have fluorescence values
below the peak value (2) if the cell's fluorescence values follow a bimodal
distribution. Each of these parameters are used in a model to predict the
likelihood of a cell being a false responder. These parameters are abberant
when the background have not been estimated correctly (eg. when fluorescence trace
has a sigmoid shape (this case is badly handled by the DPA method and creates a lot
of false positives). Background estimation with pattern matching overcomes
this problem and deals much better with tricky shapes.
}
