% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{prepareData}
\alias{prepareData}
\title{prepareData}
\usage{
prepareData(
  folder_name,
  frame_rate,
  duration_in_seconds = 30,
  marker_thresh,
  unit
)
}
\arguments{
\item{folder_name}{Path to the folder containing the fluorescence data and the meta.csv file}

\item{frame_rate}{Your Acquisition rate in Hertz (eg. 0.25 if it is 0.25 Hz, 100 if it is 100 Hz)}

\item{duration_in_seconds}{The time your stimulus takes to be in contact to your cells. Peaks found during this period will be attributed to the previous stimulus.}
}
\description{
This function is a wrapper which reads the raw fluorescence csv files (wide format),
transform it to long format and adds several meta in column (Cell_id, time, stimuli,
molecular marker, group, coverslip).
}
