% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepareData_CalQuick.R
\name{prepareData}
\alias{prepareData}
\title{prepareData}
\usage{
prepareData(
  folder_name,
  stim_number,
  frame_rate,
  duration_in_seconds = 30,
  compare_groups = FALSE,
  marker_thresh
)
}
\arguments{
\item{compare_groups}{}
}
\description{
prepareData
}
