% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Stim_Var_CalQuick.R
\name{stim_var}
\alias{stim_var}
\title{stim_var}
\usage{
stim_var(data, stimuli, each, frame_rate, coverslip_id)
}
\arguments{
\item{data}{the original dataset}

\item{stimuli}{a character vector with the names of the stimuli used}

\item{each}{an integer vector with the number of repetitions for each stimulus}
}
\value{
Returns a dataframe with a new variable called stimulus and prints
the output of a control to verify that the stimuli are repeated the right
number of times
}
\description{
stim_var
}
