% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_preparation.R
\name{trackmateInput}
\alias{trackmateInput}
\title{trackmateInput}
\usage{
trackmateInput(
  file,
  stimuli,
  each,
  frame_rate,
  coverslip_id,
  id,
  group,
  duration_in_seconds
)
}
\arguments{
\item{duration_in_seconds}{}
}
\description{
trackmateInput
}
